+++ src/SpecEdu.Web/Program.cs
@@ -1,26 +1,78 @@
+using Microsoft.AspNetCore.Localization;
+using Microsoft.AspNetCore.Mvc.Razor;
+using SpecEdu.Application.Common.Interfaces;
+using SpecEdu.Infrastructure;
+using SpecEdu.Web.Logging;
+using SpecEdu.Web.Services;
+using System.Globalization;
+
 var builder = WebApplication.CreateBuilder(args);
 
+builder.Services.AddLocalization(options => options.ResourcesPath = "Resources");
 // Add services to the container.
-builder.Services.AddRazorPages();
+builder.Services.AddRazorPages().AddViewLocalization(LanguageViewLocationExpanderFormat.Suffix)
+    .AddDataAnnotationsLocalization();
+
+// Add HttpContextAccessor for CurrentUserService
+builder.Services.AddHttpContextAccessor();
+
+// Register CurrentUserService
+builder.Services.AddScoped<ICurrentUserService, CurrentUserService>();
+
+// Add Infrastructure services (DbContext, Identity, JWT, etc.)
+builder.Services.AddInfrastructure(builder.Configuration);
 
 var app = builder.Build();
+
+// Get logger for startup logging
+var logger = app.Services.GetRequiredService<ILogger<Program>>();
+
+// Localization configuration
+var supportedCultures = new[] { "cs", "en", "de" }
+    .Select(c => new CultureInfo(c))
+    .ToList();
+
+var localizationOptions = new RequestLocalizationOptions
+{
+    DefaultRequestCulture = new RequestCulture("cs"),
+    SupportedCultures = supportedCultures,
+    SupportedUICultures = supportedCultures
+};
+
+localizationOptions.RequestCultureProviders = new IRequestCultureProvider[]
+{
+    new QueryStringRequestCultureProvider(),
+    new CookieRequestCultureProvider(),
+    new AcceptLanguageHeaderRequestCultureProvider()
+};
+
+app.UseRequestLocalization(localizationOptions);
+
+// Log application startup
+Log.ApplicationStarting(logger, "0.0.8");
+Log.ConfigurationLoaded(logger, "ConnectionStrings");
+Log.ConfigurationLoaded(logger, "JwtSettings");
 
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
 {
     app.UseExceptionHandler("/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
 }
 
 app.UseHttpsRedirection();
 
+app.UseStaticFiles();
+
 app.UseRouting();
 
+app.UseAuthentication();
 app.UseAuthorization();
 
 app.MapStaticAssets();
 app.MapRazorPages()
    .WithStaticAssets();
+
+Log.ApplicationStarted(logger);
 
 app.Run();