+++ src/SpecEdu.Web/SharedResource.cs
@@ -1,0 +1,6 @@
+﻿namespace SpecEdu.Web
+{
+    public class SharedResource
+    {
+    }
+}